// packages/js/data/src/countries/action-types.ts
export enum TYPES {
	GET_LOCALES_ERROR = 'GET_LOCALES_ERROR',
	GET_LOCALES_SUCCESS = 'GET_LOCALES_SUCCESS',
	GET_COUNTRIES_ERROR = 'GET_COUNTRIES_ERROR',
	GET_COUNTRIES_SUCCESS = 'GET_COUNTRIES_SUCCESS',
	GEOLOCATION_SUCCESS = 'GEOLOCATION_SUCCESS',
	GEOLOCATION_ERROR = 'GEOLOCATION_ERROR',
	CURRENCY_SYMBOLS_SUCCESS = 'CURRENCY_SYMBOLS_SUCCESS',
	CURRENCY_SYMBOLS_ERROR = 'CURRENCY_SYMBOLS_ERROR',
}

export default TYPES;